refactor netlink module
