+++ justfile
@@ -52,8 +52,8 @@ debug dimensions="3800x200": _lbuild
   env -u I3SOCK DISPLAY=:1.0 i3-with-shmlog --config ./scripts/i3.conf
 
 # run tests in a nested dbus session so the host session isn't affected
-test:
-  dbus-run-session -- env ISTAT_TEST=1 cargo test --all
+test *args:
+  dbus-run-session -- env RUST_LOG=trace ISTAT_TEST=1 cargo test --all "$@"
 
 # `eval` this for an easy debug loop for screenshot tests
 # NOTE: requires `fd` be present, and the terminal is `kitty`
